@@ -6,7 +6,7 @@
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.4.1",
-    "react-addons-perf": "^15.3.2"
+    "react-addons-perf": "15.3.2"
   },
   "scripts": {
     "test": "npm test",
@@ -27,29 +27,33 @@
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
     "babel-relay-plugin": "^0.9.3",
-    "babel-relay-plugin-loader": "^0.9.0",
+    "babel-relay-plugin-0.9": "^0.9.4",
+    "babel-relay-plugin-loader-0.9": "^0.9.6",
     "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
     "dialog-polyfill": "^0.4.4",
     "dompurify": "^0.8.3",
-    "es6-shim": "^0.35.1",
     "expose-loader": "^0.7.1",
     "extract-text-webpack-plugin": "^1.0.1",
-    "formsy-material-ui": "^0.5.2",
+    "formsy-material-ui": "0.5.2",
     "formsy-react": "^0.18.1",
+    "highlight.js": "^9.8.0",
+    "immutability-helper": "^2.0.0",
     "imports-loader": "^0.6.5",
     "jquery": "^3.1.0",
     "jquery-ujs": "^1.2.2",
-    "material-ui": "^0.15.0",
+    "json-loader": "^0.5.4",
+    "material-ui": "^0.16.4",
     "moment": "^2.14.1",
-    "node-sass": "^3.11.1",
+    "node-sass": "^3.12.5",
+    "promise-window": "^1.1.0",
     "query-string": "^4.2.3",
-    "react": "^15.3.2",
-    "react-addons-pure-render-mixin": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.4.0",
+    "react-dom": "^15.4.0",
     "react-loader": "^2.4.0",
     "react-relay": "^0.9.3",
-    "react-tap-event-plugin": "^1.0.0",
+    "react-tap-event-plugin": "^2.0.1",
+    "react-truncate": "^2.0.2",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "underscore": "^1.8.3",
