@@ -3,16 +3,21 @@ import React from 'react';
 import Relay from 'react-relay';
 
 const Company = props => (
-  <div>
+  <div
+    className="company"
+    style={{
+      color: 'rgba(0, 0, 0, 0.7)',
+      fontWeight: 500,
+      whiteSpace: 'nowrap',
+      width: 200,
+      textOverflow: 'ellipsis',
+      overflow: 'hidden',
+    }}
+  >
     {props.developer.company ?
-      <div
-        className="location"
-        style={{ color: 'rgba(0, 0, 0, 0.7)', fontWeight: 500, marginTop: 5 }}
-      >
-        <small className="company">
-          {props.developer.company}
-        </small>
-      </div> : ''
+      <small className="company">
+        {props.developer.company}
+      </small> : ''
     }
   </div>
 );
