@@ -4,11 +4,12 @@
 import React, { Component } from 'react';
 import Relay from 'react-relay';
 import Formsy from 'formsy-react';
+import Chip from 'material-ui/Chip';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import RaisedButton from 'material-ui/RaisedButton';
 import ActionDelete from 'material-ui/svg-icons/action/delete';
-import Chip from 'material-ui/Chip';
 import Snackbar from 'material-ui/Snackbar';
+import AutoComplete from 'material-ui/AutoComplete';
 import {
   Card,
   CardTitle,
@@ -19,6 +20,7 @@ import { css } from 'aphrodite';
 
 // Child Components
 import muiTheme from '../theme.es6';
+import LanguageList from '../../utils/languages.json';
 
 // Mutations
 import UpdateDeveloper from '../../mutations/developer/updateDeveloper.es6';
@@ -44,17 +46,18 @@ class DeveloperEdit extends Component {
 
   constructor(props) {
     super(props);
-    this.renderChip = this.renderChip.bind(this);
-    this.handleRequestDelete = this.handleRequestDelete.bind(this);
     this.handleRequestClose = this.handleRequestClose.bind(this);
     this.clearValidationErrors = this.clearValidationErrors.bind(this);
     this.submitForm = this.submitForm.bind(this);
-    this.addNewPlatform = this.addNewPlatform.bind(this);
     this.enableButton = this.enableButton.bind(this);
     this.disableButton = this.disableButton.bind(this);
     this.setNotification = this.setNotification.bind(this);
+    this.renderChip = this.renderChip.bind(this);
+    this.handleRequestDelete = this.handleRequestDelete.bind(this);
+    this.addNewLanguage = this.addNewLanguage.bind(this);
+    this.onInvalidLanguageSelection = this.onInvalidLanguageSelection.bind(this);
 
-    const platforms = props.developer.platforms.map((elem, index) => (
+    const languages = props.developer.languages.map((elem, index) => (
       { key: index, label: elem }
     ));
 
@@ -63,10 +66,16 @@ class DeveloperEdit extends Component {
       value: 1,
       loaded: false,
       canSubmit: false,
+      languages,
       notification: '',
-      platforms,
       validationErrors: {},
     };
+  }
+
+  onInvalidLanguageSelection(message) {
+    this.languageNode.state.searchText = '';
+    this.languageNode.focus();
+    this.setNotification(message);
   }
 
   setNotification(notification) {
@@ -79,7 +88,7 @@ class DeveloperEdit extends Component {
 
   submitForm(event) {
     event.preventDefault();
-    const platforms = this.state.platforms.map(elem => (
+    const languages = this.state.languages.map(elem => (
       elem.label
     ));
 
@@ -103,7 +112,7 @@ class DeveloperEdit extends Component {
     };
 
     const newModel = Object.assign(this.formNode.getModel(), {
-      platforms: platforms.toString(),
+      languages: languages.toString(),
     });
 
     Relay.Store.commitUpdate(new UpdateDeveloper({
@@ -130,75 +139,54 @@ class DeveloperEdit extends Component {
     });
   }
 
-  addNewPlatform(event) {
-    if (event.keyCode === 13) {
-      event.preventDefault();
-    }
-
-    if (event.keyCode === 13 || event.keyCode === 188 || event.keyCode === 32) {
-      const newPlatform = event.target.value.trim();
-      if (newPlatform === '') {
-        this.setState({
-          validationErrors: {
-            platforms: 'Empty language',
-          },
-        });
-
-        setTimeout(() => {
-          this.clearValidationErrors();
-        }, 3000);
-
-        return;
-      }
-
-      const isDuplicate = this.state.platforms.some(platform => (
-        platform.label === newPlatform
-      ));
-
-      if (isDuplicate) {
-        this.platformNode.state.value = '';
-
-        this.setState({
-          validationErrors: {
-            platforms: 'Duplicate platform',
-          },
-        });
-
-        setTimeout(() => {
-          this.clearValidationErrors();
-        }, 3000);
-
-        return;
-      }
-
-      const platforms = this.state.platforms.concat([{
-        key: this.state.platforms.length + 1,
-        label: newPlatform.toLowerCase(),
-      }]);
-
-      this.setState({ platforms }, () => {
-        this.platformNode.state.value = '';
-        this.setState({
-          canSubmit: true,
-        });
-      });
-    }
-  }
-
   handleRequestClose() {
     this.setState({
       open: false,
     });
   }
 
+  addNewLanguage(selectedLanguage, index) {
+    const newLanguage = selectedLanguage.trim();
+    if (index === -1) {
+      this.onInvalidLanguageSelection('Please select from the list');
+      return;
+    }
+
+    if (newLanguage === '') {
+      this.onInvalidLanguageSelection('Empty language');
+      return;
+    }
+
+    const isDuplicate = this.state.languages.some(language => (
+      language.label === newLanguage
+    ));
+
+    if (isDuplicate) {
+      this.onInvalidLanguageSelection('Duplicate language');
+      return;
+    }
+
+    const languages = this.state.languages.concat([{
+      key: this.state.languages.length + 1,
+      label: newLanguage,
+    }]);
+
+    this.setState({ languages }, () => {
+      this.languageNode.state.searchText = '';
+      this.languageNode.focus();
+      this.setState({
+        canSubmit: true,
+      });
+    });
+  }
 
   handleRequestDelete(key) {
-    this.platforms = this.state.platforms;
-    const platformToDelete = this.platforms
-    .map(platform => platform.key)
+    this.languages = this.state.languages;
+    const languageToDelete = this.languages
+    .map(language => language.key)
     .indexOf(key);
-    this.platforms.splice(platformToDelete, 1);
-    this.setState({ platforms: this.platforms });
+    this.languages.splice(languageToDelete, 1);
+    this.setState({ languages: this.languages });
   }
 
   renderChip(data) {
@@ -206,7 +194,7 @@ class DeveloperEdit extends Component {
       <Chip
         key={data.key}
         onRequestDelete={() => this.handleRequestDelete(data.key)}
-        className={css(chipStyles.chip)}
+        className={css(chipStyles.badge)}
       >
         {data.label}
       </Chip>
@@ -215,7 +203,7 @@ class DeveloperEdit extends Component {
 
   render() {
     const { developer } = this.props;
-
+    const languageKeys = Object.keys(LanguageList);
     return (
       <MuiThemeProvider muiTheme={muiTheme}>
         <Card
@@ -237,11 +225,12 @@ class DeveloperEdit extends Component {
               <Formsy.Form
                 onValid={this.enableButton}
                 autoComplete="off"
+                className="form"
                 ref={node => (this.formNode = node)}
                 onInvalid={this.disableButton}
                 validationErrors={this.state.validationErrors}
               >
-                <div className="search-box bio">
+                <div className="field bio">
                   <FormsyText
                     id="text-field-default"
                     placeholder="(ex: Looking for opportunities in AI)"
@@ -265,7 +254,7 @@ class DeveloperEdit extends Component {
                   />
                 </div>
 
-                <div className="search-box">
+                <div className="field">
                   <FormsyText
                     id="text-field-default"
                     placeholder="(ex: London, Toronto, Oslo)"
@@ -281,13 +270,13 @@ class DeveloperEdit extends Component {
 
                   <FormsyText
                     id="text-field-default"
-                    placeholder="(ex: https://linkedin.com)"
-                    name="linkedin"
+                    placeholder="(ex: https://blog.com)"
+                    name="blog"
                     className={css(formStyles.half)}
                     fullWidth
                     onKeyDown={DeveloperEdit.onKeyPress}
-                    defaultValue={developer.linkedin}
-                    floatingLabelText="Linkedin"
+                    defaultValue={developer.blog}
+                    floatingLabelText="Blog or Website"
                     floatingLabelFixed
                     updateImmediately
                     validations={{
@@ -298,24 +287,42 @@ class DeveloperEdit extends Component {
                     }}
                   />
                 </div>
-
                 <div className="search-box">
-                  <FormsyText
+                  <AutoComplete
                     id="text-field-default"
                     placeholder="(ex: ruby, python)"
-                    name="platforms"
-                    className="platforms"
-                    onKeyDown={this.addNewPlatform}
-                    ref={node => (this.platformNode = node)}
+                    name="languages"
+                    className="languages"
+                    onNewRequest={this.addNewLanguage}
+                    onKeyDown={DeveloperEdit.onKeyPress}
                     floatingLabelText="Languages and frameworks you work with *"
                     floatingLabelFixed
                     fullWidth
+                    ref={node => (this.languageNode = node)}
+                    filter={AutoComplete.fuzzyFilter}
+                    dataSource={languageKeys}
+                    maxSearchResults={5}
                   />
 
+                  {/*// <FormsyText
+                  //   id="text-field-default"
+                  //   placeholder="(ex: ruby, python)"
+                  //   name="languages"
+                  //   className="languages"
+                  //   onKeyDown={this.addNewLanguage}
+                  //   ref={node => (this.languageNode = node)}
+                  //   floatingLabelText="Languages and frameworks you work with *"
+                  //   floatingLabelFixed
+                  //   fullWidth
+                  // />
+                  */}
+
                   <div className={css(chipStyles.wrapper)}>
-                    {this.state.platforms.map(this.renderChip, this)}
+                    {this.state.languages.map(this.renderChip, this)}
                   </div>
                 </div>
+
+                <div className="clearfix" />
 
                 <div className="header-separator">Availability</div>
                 <div className="hireable">
@@ -333,7 +340,7 @@ class DeveloperEdit extends Component {
 
                 <div className="clearfix" />
                 <div className="header-separator">Preferences</div>
-                <div className="search-box">
+                <div className="field">
                   <div className={css(formStyles.preferences)}>
                     <FormsyCheckbox
                       label="Relocate"
@@ -351,7 +358,7 @@ class DeveloperEdit extends Component {
                   </div>
 
                   <div className="clearfix" />
-                  <div className="search-box job">
+                  <div className="field job">
                     <div className="header-separator">Job Type</div>
                     <FormsyCheckbox
                       label="Part-Time"
@@ -397,7 +404,7 @@ class DeveloperEdit extends Component {
                   </div>
 
                   <div className="clearfix" />
-                  <div className="search-box levels">
+                  <div className="field levels">
                     <div className="header-separator">Experience Level</div>
                     <FormsyCheckbox
                       label="CTO"
@@ -455,12 +462,20 @@ class DeveloperEdit extends Component {
                   />
 
                   <RaisedButton
+                    label="Back"
+                    primary
+                    type="submit"
+                    style={{ marginLeft: 10 }}
+                    href={Routes.root_path()}
+                  />
+
+                  <RaisedButton
                     secondary
                     label="Delete account"
                     icon={<ActionDelete />}
                     data-method="delete"
                     data-confirm="This will delete your account. Okay?"
-                    className="edit-link pull-right"
+                    className="delete-link pull-right"
                     href={Routes.cancel_developer_registration_path()}
                   />
                 </div>
@@ -495,10 +510,9 @@ const DeveloperEditContainer = Relay.createContainer(DeveloperEdit, {
         name,
         login,
         bio,
-        linkedin,
+        blog,
         location,
-
-        platforms,
+        languages,
 
         hireable,
 
