@@ -6,8 +6,8 @@ import FontIcon from 'material-ui/FontIcon';
 import Chip from 'material-ui/Chip';
 import Avatar from 'material-ui/Avatar';
 
-// StyleSheets
-import chipStyles from '../styles/chips.es6';
+// Stylesheet
+import iconStyles from '../styles/icons.es6';
 
 const JobTypes = (props) => {
   const { developer } = props;
@@ -20,15 +20,20 @@ const JobTypes = (props) => {
         developer.freelance ||
         developer.startup ||
         developer.internship ?
-          <div className={css(chipStyles.wrapper)}>
-            <div className="header-separator">Job Types</div>
+          <div className={css(iconStyles.links)}>
+            <div className="header-separator">Job Preferences</div>
             {developer.full_time ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Full-Time
               </Chip> : ''
@@ -36,11 +41,16 @@ const JobTypes = (props) => {
 
             {developer.part_time ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Part-Time
               </Chip> : ''
@@ -48,11 +58,16 @@ const JobTypes = (props) => {
 
             {developer.contract ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Contract
               </Chip> : ''
@@ -61,11 +76,16 @@ const JobTypes = (props) => {
 
             {developer.freelance ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Freelance
               </Chip> : ''
@@ -73,11 +93,16 @@ const JobTypes = (props) => {
 
             {developer.startup ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Startup
               </Chip> : ''
@@ -85,11 +110,16 @@ const JobTypes = (props) => {
 
             {developer.internship ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">work</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >work</FontIcon>}
                 />
                 Internship
               </Chip> : ''
