@@ -6,8 +6,8 @@ import FontIcon from 'material-ui/FontIcon';
 import Chip from 'material-ui/Chip';
 import Avatar from 'material-ui/Avatar';
 
-// StyleSheets
-import chipStyles from '../styles/chips.es6';
+// Stylesheet
+import iconStyles from '../styles/icons.es6';
 
 const Levels = (props) => {
   const { developer } = props;
@@ -20,15 +20,20 @@ const Levels = (props) => {
         developer.mid ||
         developer.junior ||
         developer.student ?
-          <div className={css(chipStyles.wrapper)}>
-            <div className="header-separator">Levels</div>
+          <div className={css(iconStyles.links)}>
             {developer.cto ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 CTO
               </Chip> : ''
@@ -36,11 +41,17 @@ const Levels = (props) => {
 
             {developer.lead ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 Lead
               </Chip> : ''
@@ -48,11 +59,17 @@ const Levels = (props) => {
 
             {developer.senior ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 Senior
               </Chip> : ''
@@ -61,11 +78,17 @@ const Levels = (props) => {
 
             {developer.mid ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 Mid-Level
               </Chip> : ''
@@ -73,11 +96,17 @@ const Levels = (props) => {
 
             {developer.junior ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 Junior
               </Chip> : ''
@@ -85,11 +114,17 @@ const Levels = (props) => {
 
             {developer.student ?
               <Chip
+                labelStyle={{ fontSize: 14 }}
                 key={Math.random()}
-                className={css(chipStyles.chip)}
+                className={css(iconStyles.linkIcon)}
               >
                 <Avatar
-                  icon={<FontIcon className="material-icons">label</FontIcon>}
+                  className={css(iconStyles.iconAvatar)}
+                  icon={
+                    <FontIcon
+                      className={`material-icons ${css(iconStyles.chipIcon)}`}
+                    >label</FontIcon>
+                  }
                 />
                 Student
               </Chip> : ''
