@@ -5,7 +5,16 @@ import Relay from 'react-relay';
 
 const Name = props => (
   <div className="name">
-    <h3 style={{ margin: 0, fontWeight: '500' }}>
+    <h3
+      style={{
+        margin: 0,
+        fontWeight: '500',
+        whiteSpace: 'nowrap',
+        width: 260,
+        textOverflow: 'ellipsis',
+        overflow: 'hidden',
+      }}
+    >
       {props.developer.name ? props.developer.name : 'Anonymous'}
     </h3>
   </div>
