@@ -2,23 +2,29 @@
 import React from 'react';
 import Relay from 'react-relay';
 import Avatar from 'material-ui/Avatar';
-import Chip from 'material-ui/Chip';
 
 const Orgs = props => (
   <div>
     {props.developer.orgs.length > 0 ?
-      <div className="orgs" style={{ display: 'flex', flexWrap: 'wrap' }}>
-        <div className="header-separator">Organizations</div>
+      <div
+        className="orgs"
+        style={{
+          display: 'flex',
+          flexWrap: 'wrap',
+          justifyContent: 'center',
+        }}
+      >
         {props.developer.orgs.map(org => (
           <div
             className="org"
             key={org.id}
             style={{ marginRight: 10, marginBottom: 10 }}
           >
-            <Chip>
-              <Avatar src={org.avatar_url} size={50} />
-              {org.login}
-            </Chip>
+            <Avatar
+              src={org.avatar_url}
+              size={32}
+              style={{ borderRadius: 0 }}
+            />
           </div>
         ))}
       </div> : ''
