@@ -5,9 +5,7 @@ import React from 'react';
 import Relay from 'react-relay';
 import Avatar from 'material-ui/Avatar';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import RaisedButton from 'material-ui/RaisedButton';
 import { css } from 'aphrodite';
-import ActionEdit from 'material-ui/svg-icons/image/edit';
 
 // Child components
 import Languages from './languages.es6';
@@ -22,9 +20,10 @@ import Bio from './bio.es6';
 import Orgs from './orgs.es6';
 import Company from './company.es6';
 import Actions from './actions.es6';
+import Achievements from './achievements.es6';
+import Preferences from './preferences.es6';
 
 // StyleSheets
-import badgeStyles from '../styles/badges.es6';
 import '../styles/profile.sass';
 
 // Utils
@@ -41,56 +40,51 @@ const DeveloperShow = (props) => {
         <header className="header header--bg">
           <div className="container">
             <div className="wrapper">
-              <div
-                className="image"
-              >
+              <div className="image">
                 {developer.hireable ?
-                  <div
-                    className={
-                      css(
-                        badgeStyles.badge,
-                        badgeStyles.hireable,
-                        badgeStyles.big
-                      )
-                    }
-                  > H </div> : ''
+                  <div className="badge hireable big"> H
+                  <span className="full">ireable</span></div> :
+                  !developer.company ?
+                    <div className="badge hireable big maybe"> M
+                    <span className="full">ay be hireable</span></div> : ''
                 }
                 <Avatar
                   src={developer.avatar_url}
                   size={150}
+                  className="avatar"
                 />
+
                 <br />
-
-                <Meta developer={developer} />
                 {currentUser.isEmployer ? <Actions developer={developer} /> : ''}
-
-                {props.can_edit ?
-                  <RaisedButton
-                    primary
-                    label="Edit"
-                    icon={<ActionEdit />}
-                    className="edit-link"
-                    href={Routes.edit_developer_path(developer.login)}
-                  /> : ''
+                <Meta developer={developer} />
+                <Orgs developer={developer} />
+              </div>
+              <div className="basic">
+                <Name developer={developer} />
+                <div style={{ display: 'flex' }}>
+                  <Location developer={developer} />
+                  <Company developer={developer} />
+                </div>
+                <Bio developer={developer} />
+                <Links developer={developer} signedIn={props.signedIn} />
+                <Languages developer={developer} />
+                {props.signedIn ?
+                  <div className="pro-info">
+                    <JobTypes developer={developer} />
+                    <Preferences developer={developer} />
+                    <Levels developer={developer} />
+                  </div> : ''
                 }
               </div>
 
-              <div className="profile">
-                <div className="basic">
-                  <Name developer={developer} />
-                  <Location developer={developer} />
-                  <Company developer={developer} />
-                  <Bio developer={developer} />
-                </div>
-                <Links developer={developer} />
-                <div className="clearfix" />
-                <Orgs developer={developer} />
-                <Languages developer={developer} />
-                <JobTypes developer={developer} />
-                <Levels developer={developer} />
-              </div>
+              {developer.premium ?
+                <div className="premium-badge">
+                  Premium
+                </div> : ''
+              }
             </div>
           </div>
+          <Achievements developer={developer} />
         </header>
       </div>
     </MuiThemeProvider>
@@ -98,8 +92,8 @@ const DeveloperShow = (props) => {
 };
 
 DeveloperShow.propTypes = {
-  can_edit: React.PropTypes.bool,
   developer: React.PropTypes.object,
+  signedIn: React.PropTypes.bool,
 };
 
 const DeveloperShowContainer = Relay.createContainer(DeveloperShow, {
@@ -111,6 +105,9 @@ const DeveloperShowContainer = Relay.createContainer(DeveloperShow, {
         login,
         hireable,
         premium,
+        company,
+        ${Preferences.getFragment('developer')}
+        ${Achievements.getFragment('developer')},
         ${Name.getFragment('developer')},
         ${Company.getFragment('developer')},
         ${Orgs.getFragment('developer')},
