@@ -24,7 +24,9 @@ const LoadingList = () => {
     >
       <ListItem
         leftAvatar={
-          <Avatar src={'https://placeholdit.imgix.net/~text?w=80&h=80'} size={80} />
+          <div style={{ textAlign: 'center' }}>
+            <Avatar src={'https://placeholdit.imgix.net/~text?w=80&h=80'} size={80} />
+          </div>
         }
         secondaryText={
           <div className="animated-background secondary" />
@@ -40,7 +42,13 @@ const LoadingList = () => {
 
   return (
     <MuiThemeProvider>
-      <Card>
+      <Card
+        style={{
+          boxShadow: 'none',
+          border: 0,
+          borderRadius: 0,
+        }}
+      >
         <CardText style={{ padding: 0, overflow: 'hidden' }}>
           <List style={{ paddingTop: '0px', paddingBottom: '0px' }}>
             {emptyPlaceholders}
