@@ -9,20 +9,28 @@ const badges = StyleSheet.create({
   },
 
   big: {
-    left: 160,
+    top: 20,
+    left: 130,
   },
 
   tag: {
     border: '1px solid #777',
     display: 'inline-block',
     padding: '5px 10px',
+    cursor: 'pointer',
     marginRight: 10,
     marginBottom: 10,
     textDecoration: 'none',
     color: '#333',
     minWidth: 50,
-    letterSpacing: '0.02em',
     fontSize: 12,
+    letterSpacing: '0.02em',
+  },
+
+  hover: {
+    ':hover': {
+      backgroundColor: '#f2f2f2',
+    },
   },
 
   hireable: {
@@ -38,8 +46,22 @@ const badges = StyleSheet.create({
     left: 65,
     top: 6,
     zIndex: 10,
+    overflow: 'hidden',
+    display: 'block',
+    transition: 'all 0.5s ease',
     boxShadow: 'rgba(63, 67, 69, 0.298039) 0px 0px 16px 0px',
     border: '1px solid white',
+  },
+
+  full: {
+    display: 'none',
+  },
+
+  expand: {
+    ':hover': {
+      width: '100px',
+      borderRadius: '5px',
+    },
   },
 });
 
