@@ -3,7 +3,6 @@ import { StyleSheet } from 'aphrodite';
 const iconStyles = StyleSheet.create({
   links: {
     height: 'auto',
-    marginTop: 15,
     marginLeft: 0,
     display: 'flex',
     flexWrap: 'wrap',
@@ -12,14 +11,40 @@ const iconStyles = StyleSheet.create({
   linkIcon: {
     color: '#555',
     padding: 0,
-    marginRight: 10,
-    marginBottom: 5,
+    borderRadius: 0,
+    fontSize: 13,
+    backgroundColor: 'transparent',
+  },
+
+  bordered: {
+    margin: 5,
+    border: '1px solid #f2f2f2',
+  },
+
+  premium: {
+    border: '1px solid #fff',
+  },
+
+  hover: {
+    ':hover': {
+      backgroundColor: '#f2f2f2',
+    },
+  },
+
+  label: {
+    fontSize: 16,
   },
 
   iconAvatar: {
-    backgroundColor: 'rgb(91, 152, 224)',
+    backgroundColor: 'transparent',
+    color: 'rgb(91, 152, 224)',
+    fontSize: 16,
   },
 
+  chipIcon: {
+    backgroundColor: 'transparent',
+    color: 'rgb(91, 152, 224)',
+  },
 });
 
 export default iconStyles;
