@@ -28,7 +28,10 @@ class ApplicationController < ActionController::Base
   def configure_permitted_parameters
     devise_parameter_sanitizer.permit(
       :sign_up,
-      keys: [:name, :company, :website, :language, :location]
+      keys: [
+        :name, :company, :website,
+        preferences: [:location, :language, :type]
+      ]
     )
   end
 
