@@ -3,7 +3,6 @@ module GraphqlAuthentication
 
   included do
     before_action :find_current_employer, :find_current_developer
-    before_action :render_unauthorised, unless: :authenticated?
   end
 
   def authenticated?
@@ -15,14 +14,16 @@ module GraphqlAuthentication
   end
 
   def find_current_employer
+    return if cookies.signed['employer.login'].nil?
     @find_current_employer ||= Employer.find_by_login(
       cookies.signed['employer.login']
-    ) unless cookies.signed['employer.login'].nil?
+    )
   end
 
   def find_current_developer
+    return if cookies.signed['developer.login'].nil?
     @find_current_developer ||= Developer.find_by_login(
       cookies.signed['developer.login']
-    ) unless cookies.signed['developer.login'].nil?
+    )
   end
 end
