@@ -1,15 +1,15 @@
 Schema = GraphQL::Schema.define do
   query QueryType
   mutation MutationType
-  max_depth 15
+  max_depth 8
   rescue_from ActiveRecord::RecordInvalid, &:messag
   rescue_from ActiveRecord::Rollback, &:message
   rescue_from StandardError, &:message
   rescue_from ActiveRecord::RecordNotUnique, &:message
   rescue_from ActiveRecord::RecordNotFound, &:message
-  object_from_id -> (id, _ctx) { decode_object(id) }
-  id_from_object -> (obj, type, _ctx) { encode_object(obj, type) }
-  resolve_type -> (object, _ctx) { Schema.types[type_name(object)] }
+  object_from_id ->(id, _ctx) { decode_object(id) }
+  id_from_object ->(obj, type, _ctx) { encode_object(obj, type) }
+  resolve_type ->(object, _ctx) { Schema.types[type_name(object)] }
 end
 
 def type_name(object)
@@ -19,7 +19,7 @@ end
 def encode_object(object, type)
   GraphQL::Schema::UniqueWithinType.encode(
     type.name,
-    object.login,
+    object.id,
     separator: '---'
   )
 end
@@ -29,7 +29,7 @@ def decode_object(id)
     id,
     separator: '---'
   )
-  Object.const_get(type_name).find_by_login(object_id)
+  Object.const_get(type_name).find(object_id)
 end
 
 # Responsible for dumping Schema.json
@@ -56,10 +56,8 @@ module SchemaHelpers
     # Generate the schema on start/reload
     FileUtils.mkdir_p SCHEMA_DIR
     result = JSON.pretty_generate(Schema.execute_introspection_query)
-    File.write(
-      SCHEMA_PATH,
-      result
-    ) unless File.exist?(SCHEMA_PATH) && File.read(SCHEMA_PATH) == result
+    return if File.exist?(SCHEMA_PATH) && File.read(SCHEMA_PATH) == result
+    File.write(SCHEMA_PATH, result)
   end
 end
 
