@@ -5,20 +5,20 @@ QueryType = GraphQL::ObjectType.define do
 
   field :root, ViewerType do
     description 'Root field to query related collections/objects'
-    resolve -> (_obj, _args, _ctx) { Viewer::STATIC }
+    resolve ->(_obj, _args, _ctx) { Viewer::STATIC }
   end
 
   field :developer do
     argument :id, !types.String
     type DeveloperType
     description 'Returns a developer profile by id'
-    resolve(DeveloperResolver)
+    resolve(Developers::ShowResolver)
   end
 
   field :employer do
     argument :id, !types.String
     type EmployerType
     description 'Returns a employer profile by id'
-    resolve(EmployerResolver)
+    resolve(Employers::ShowResolver)
   end
 end
