@@ -13,7 +13,12 @@ class PrepareSearchParams
   end
 
   def to_props
-    search_props = { first: 50, list: true, page: Integer(params[:page] || 1) }
+    search_props = {
+      signedIn: current_user.present?,
+      first: 50,
+      list: true,
+      page: Integer(params[:page] || 1)
+    }
     supported.each do |key|
       search_props[key] = params[key.to_sym]
     end
