@@ -1,8 +1,8 @@
-Id = Struct.new :login
+Id = Struct.new :id
 class Viewer < Id
-  STATIC = new(login: 'root').freeze
+  STATIC = new(id: 'root').freeze
 
-  def self.find_by_login(_)
+  def self.find(_)
     STATIC
   end
 end
