@@ -1,7 +1,7 @@
 # Use this hook to configure devise mailer, warden hooks and so forth.
 # Many of these configuration options can be set straight in your model.
 Devise.setup do |config|
-  config.mailer_sender = 'noreply@hireables.co'
+  config.mailer_sender = 'signups@hireables.co'
 
   # Configure the class responsible to send e-mails.
   # config.mailer = 'Devise::Mailer'
@@ -239,6 +239,16 @@ Devise.setup do |config|
   # config.omniauth :github, 'APP_ID', 'APP_SECRET', scope: 'user,public_repo'
 
   config.omniauth :github, ENV['GITHUB_KEY'], ENV['GITHUB_SECRET'], scope: 'user, public_repo, read:org'
+  config.omniauth :producthunt, ENV['PRODUCTHUNT_API_KEY'], ENV['PRODUCTHUNT_API_SECRET'], callback_url: "#{ENV['SITE_DOMAIN']}/developers/auth/producthunt/callback"
+  config.omniauth :meetup, ENV['MEETUP_API_KEY'], ENV['MEETUP_API_SECRET'], callback_url: "#{ENV['SITE_DOMAIN']}/developers/auth/meetup/callback"
+  config.omniauth :stackexchange, ENV['STACKOVERFLOW_CLIENT_ID'], ENV['STACKOVERFLOW_CLIENT_SECRET'], public_key: ENV['STACKOVERFLOW_CLIENT_KEY'], callback_url: "#{ENV['SITE_DOMAIN']}/developers/auth/stackexchange/callback"
+  config.omniauth :linkedin, ENV['LINKEDIN_CLIENT_ID'], ENV['LINKEDIN_CLIENT_SECRET'], callback_url: "#{ENV['SITE_DOMAIN']}/developers/auth/linkedin/callback"
+  config.omniauth :google_oauth2, ENV['GOOGLE_CLIENT_ID'], ENV['GOOGLE_CLIENT_SECRET'], {
+    name: 'google',
+    scope: "email, profile, https://www.googleapis.com/auth/youtube",
+    prompt: 'select_account',
+    redirect_uri: "#{ENV['SITE_DOMAIN']}/developers/auth/google/callback"
+  }
 
   # ==> Warden configuration
   # If you want to use other strategies, that are not supported by Devise, or
