@@ -10,49 +10,61 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161105044438) do
+ActiveRecord::Schema.define(version: 20161115141410) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
   enable_extension "citext"
 
+  create_table "connections", force: :cascade do |t|
+    t.string   "uid"
+    t.string   "provider"
+    t.boolean  "importing",    default: false
+    t.string   "access_token"
+    t.datetime "expires_at"
+    t.integer  "developer_id"
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
+    t.index ["developer_id"], name: "index_connections_on_developer_id", using: :btree
+    t.index ["uid", "provider"], name: "index_connections_on_uid_and_provider", unique: true, using: :btree
+  end
+
   create_table "developers", force: :cascade do |t|
-    t.string   "login",              default: "",    null: false
-    t.string   "provider",           default: "",    null: false
-    t.bigint   "uid",                default: 0,     null: false
-    t.string   "avatar",             default: ""
-    t.string   "name",               default: "",    null: false
-    t.string   "email",              default: "",    null: false
+    t.string   "login",               default: "",    null: false
+    t.string   "avatar",              default: ""
+    t.string   "name",                default: "",    null: false
+    t.string   "email",               default: "",    null: false
     t.text     "bio"
-    t.string   "linkedin",           default: ""
-    t.string   "location",           default: ""
-    t.string   "platforms",          default: [],                 array: true
-    t.boolean  "remote",             default: false
-    t.boolean  "relocate",           default: false
-    t.boolean  "hireable",           default: false
-    t.boolean  "premium",            default: false
-    t.boolean  "part_time",          default: false
-    t.boolean  "full_time",          default: false
-    t.boolean  "contract",           default: false
-    t.boolean  "freelance",          default: false
-    t.boolean  "internship",         default: false
-    t.boolean  "startup",            default: false
-    t.boolean  "cto",                default: false
-    t.boolean  "lead",               default: false
-    t.boolean  "senior",             default: false
-    t.boolean  "mid",                default: false
-    t.boolean  "junior",             default: false
-    t.boolean  "student",            default: false
-    t.string   "access_token",       default: "",    null: false
-    t.string   "encrypted_password", default: "",    null: false
-    t.jsonb    "data",               default: "{}",  null: false
-    t.integer  "sign_in_count",      default: 0,     null: false
+    t.string   "linkedin",            default: ""
+    t.string   "location",            default: ""
+    t.string   "languages",           default: [],                 array: true
+    t.boolean  "remote",              default: false
+    t.boolean  "relocate",            default: false
+    t.boolean  "hireable",            default: false
+    t.boolean  "premium",             default: false
+    t.boolean  "part_time",           default: false
+    t.boolean  "full_time",           default: false
+    t.boolean  "contract",            default: false
+    t.boolean  "freelance",           default: false
+    t.boolean  "internship",          default: false
+    t.boolean  "startup",             default: false
+    t.boolean  "cto",                 default: false
+    t.boolean  "lead",                default: false
+    t.boolean  "senior",              default: false
+    t.boolean  "mid",                 default: false
+    t.boolean  "junior",              default: false
+    t.boolean  "student",             default: false
+    t.string   "encrypted_password",  default: "",    null: false
+    t.jsonb    "data",                default: "{}",  null: false
+    t.integer  "sign_in_count",       default: 0,     null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.inet     "current_sign_in_ip"
     t.inet     "last_sign_in_ip"
-    t.datetime "created_at",                         null: false
-    t.datetime "updated_at",                         null: false
+    t.datetime "created_at",                          null: false
+    t.datetime "updated_at",                          null: false
+    t.datetime "remember_created_at"
+    t.string   "blog"
     t.index ["contract"], name: "index_developers_on_contract", using: :btree
     t.index ["cto"], name: "index_developers_on_cto", using: :btree
     t.index ["data"], name: "index_developers_on_data", using: :gin
@@ -62,19 +74,18 @@ ActiveRecord::Schema.define(version: 20161105044438) do
     t.index ["hireable"], name: "index_developers_on_hireable", using: :btree
     t.index ["internship"], name: "index_developers_on_internship", using: :btree
     t.index ["junior"], name: "index_developers_on_junior", using: :btree
+    t.index ["languages"], name: "index_developers_on_languages", using: :gin
     t.index ["lead"], name: "index_developers_on_lead", using: :btree
     t.index ["location"], name: "index_developers_on_location", using: :btree
     t.index ["login"], name: "index_developers_on_login", unique: true, using: :btree
     t.index ["mid"], name: "index_developers_on_mid", using: :btree
     t.index ["part_time"], name: "index_developers_on_part_time", using: :btree
-    t.index ["platforms"], name: "index_developers_on_platforms", using: :gin
     t.index ["premium"], name: "index_developers_on_premium", using: :btree
     t.index ["relocate"], name: "index_developers_on_relocate", using: :btree
     t.index ["remote"], name: "index_developers_on_remote", using: :btree
     t.index ["senior"], name: "index_developers_on_senior", using: :btree
     t.index ["startup"], name: "index_developers_on_startup", using: :btree
     t.index ["student"], name: "index_developers_on_student", using: :btree
-    t.index ["uid"], name: "index_developers_on_uid", unique: true, using: :btree
   end
 
   create_table "employers", force: :cascade do |t|
@@ -116,6 +127,38 @@ ActiveRecord::Schema.define(version: 20161105044438) do
     t.index ["login"], name: "index_favourites_on_login", using: :btree
   end
 
+  create_table "imports", force: :cascade do |t|
+    t.string   "source_id"
+    t.string   "source_name"
+    t.jsonb    "data"
+    t.boolean  "pinned",        default: false
+    t.integer  "connection_id"
+    t.integer  "developer_id"
+    t.datetime "created_at",                    null: false
+    t.datetime "updated_at",                    null: false
+    t.index ["connection_id"], name: "index_imports_on_connection_id", using: :btree
+    t.index ["data"], name: "index_imports_on_data", using: :gin
+    t.index ["developer_id"], name: "index_imports_on_developer_id", using: :btree
+    t.index ["pinned"], name: "achievements", where: "(pinned = true)", using: :btree
+    t.index ["pinned"], name: "index_imports_on_pinned", using: :btree
+    t.index ["source_id"], name: "index_imports_on_source_id", using: :btree
+    t.index ["source_name", "source_id", "connection_id"], name: "index_imports_on_source_name_and_source_id_and_connection_id", unique: true, using: :btree
+  end
+
+  create_table "que_jobs", primary_key: ["queue", "priority", "run_at", "job_id"], force: :cascade, comment: "3" do |t|
+    t.integer   "priority",    limit: 2, default: 100,            null: false
+    t.datetime  "run_at",                default: -> { "now()" }, null: false
+    t.bigserial "job_id",                                         null: false
+    t.text      "job_class",                                      null: false
+    t.json      "args",                  default: [],             null: false
+    t.integer   "error_count",           default: 0,              null: false
+    t.text      "last_error"
+    t.text      "queue",                 default: "",             null: false
+  end
+
+  add_foreign_key "connections", "developers"
   add_foreign_key "favourites", "developers"
   add_foreign_key "favourites", "employers"
+  add_foreign_key "imports", "connections"
+  add_foreign_key "imports", "developers"
 end
